version 3

+++ assign4.pde
@@ -127,8 +127,8 @@ void keyPressed() {
 /*---------Make Alien Function-------------*/
 void alienMaker(float alienNum, float inputCol) {
   for (int i =0; i<alienNum; i++) {
-    int col = i % inputCol;
-    int row = i / inputCol;
+    float col = i % inputCol;
+    float row = i / inputCol;
     aList[i] = new Alien(50 + col*40, 50 + row*50);
   }
 }
